+++ scripts/filter_fastq_by_readids_highmem_pe.py
@@ -1,6 +1,5 @@
 #!/usr/bin/env python
 import HTSeq
-import sys
 import argparse
 import os