+++ src/ccbr_tools/pipeline/hpc.py
@@ -8,7 +8,6 @@ which contains default attributes for supported clusters.
 import pathlib
 import re
 import shutil
-import subprocess
 
 from .cache import get_singularity_cachedir, get_sif_cache_dir
 from ..shell import shell_run
@@ -25,6 +24,8 @@ class Cluster:
         singularity_sif_dir (str): The directory where Singularity SIF files are stored.
         env_vars (str): A string representing the environment variables to be set on the cluster.
     """
+
+    SPOOK_DIR = pathlib.Path("/tmp/spook")
 
     def __init__(self):
         self.name = None
@@ -79,7 +80,7 @@ class Biowulf(Cluster):
     GROUP = "CCBR_Pipeliner"
     PIPELINES_HOME = pathlib.Path("/data/CCBR_Pipeliner/Pipelines")
     TOOLS_HOME = pathlib.Path("/data/CCBR_Pipeliner/Tools")
-    CONDA_ACTIVATE = ". '/data/CCBR_Pipeliner/db/PipeDB/Conda2025/etc/profile.d/conda.sh' && conda activate /data/CCBR_Pipeliner/db/PipeDB/Conda2025/envs/py3.11-8"
+    CONDA_ACTIVATE = "MAMBA_ROOT_PREFIX='/data/CCBR_Pipeliner/db/PipeDB/miniforge3/' && mamba activate /data/CCBR_Pipeliner/db/PipeDB/miniforge3/envs/py3.11-8"
     SPOOK_DIR = pathlib.Path("/data/CCBR_Pipeliner/userdata_staging")
 
     def __init__(self):
@@ -118,7 +119,7 @@ class FRCE(Cluster):
     GROUP = "nci-frederick-ccbr-pipelines"
     PIPELINES_HOME = pathlib.Path("/mnt/projects/CCBR-Pipelines/pipelines")
     TOOLS_HOME = pathlib.Path("/mnt/projects/CCBR-Pipelines/tools")
-    CONDA_ACTIVATE = ". '/mnt/projects/CCBR-Pipelines/resources/miniconda3/etc/profile.d/conda.sh' && conda activate py311"
+    CONDA_ACTIVATE = "MAMBA_ROOT_PREFIX='/mnt/projects/CCBR-Pipelines/resources/miniconda3/' && mamba activate py311"
     SPOOK_DIR = pathlib.Path(
         "/mnt/projects/CCBR-Pipelines/pipelines/userdata/ccbrpipeliner"
     )