+++ src/ccbr_tools/pipeline/nextflow.py
@@ -13,7 +13,6 @@ import pathlib
 from ..pkg_util import msg_box
 from ..shell import shell_run
 from ..templates import use_template
-from .cache import get_singularity_cachedir
 from .util import copy_config
 from .hpc import get_hpc
 
@@ -36,6 +35,8 @@ def run(
     debug=False,
     hpc=get_hpc(),
     hpc_modules="nextflow",
+    hpc_walltime="1-00:00:00",
+    hpc_memory="1G",
 ):
     """
     Runs a Nextflow workflow with support for local or SLURM (HPC) execution modes.
@@ -49,6 +50,8 @@ def run(
         debug (bool, optional): If True, prints commands without executing them. Defaults to False.
         hpc (object, optional): HPC environment object, used for SLURM execution and module loading. Defaults to result of `~ccbr_tools.pipeline.hpc.get_hpc()`.
         hpc_modules (str, optional): Name(s) of modules to load for Nextflow execution on HPC. Defaults to "nextflow".
+        hpc_walltime (str, optional): Walltime for SLURM job submission. Defaults to "1-00:00:00".
+        hpc_memory (str, optional): Memory allocation for SLURM job submission. Defaults to "1G".
 
     Behavior:
         - Constructs the Nextflow command with appropriate arguments and profiles.
@@ -88,9 +91,7 @@ def run(
         profiles.add("slurm")
     if hpc:
         profiles.add(hpc.name)
-    if (
-        profiles
-    ):  # only add to the profiles if there are any. there are none when champagne is run on GitHub Actions.
+    if profiles:  # only add to the profiles if there are any. there are none when champagne is run on GitHub Actions.
         args_dict["-profile"] = ",".join(sorted(profiles))
 
     # use -resume by default, or do not use resume if force_all is True
@@ -116,7 +117,10 @@ def run(
     if mode == "slurm":
         slurm_filename = "submit_slurm.sh"
         use_template(
-            "submit_slurm.sh",
+            slurm_filename,
+            output_filepath=slurm_filename,
+            MEMORY=hpc_memory,
+            WALLTIME=hpc_walltime,
             PIPELINE=pipeline_name,
             MODULES=hpc_modules,
             ENV_VARS="\n".join(