+++ src/ccbr_tools/templates/__init__.py
@@ -68,7 +68,7 @@ def use_template(template_name, output_filepath=None, **kwargs: str):
         IOError: If there is an error writing the output file.
 
     Examples:
-       >>> use_template("submit_slurm.sh", output_filepath="./submit_slurm.sh", PIPELINE="CCBR_nxf", MODULES="ccbrpipeliner nextflow", ENV_VARS="", RUN_COMMAND="nextflow run main.nf -stub")
+       >>> use_template("submit_slurm.sh", output_filepath="./submit_slurm.sh", MEMORY='1G', WALLTIME="1-00:00:00", PIPELINE="CCBR_nxf", MODULES="ccbrpipeliner nextflow", ENV_VARS="", RUN_COMMAND="nextflow run main.nf -stub")
     """
     template_str = read_template(template_name)
     if not output_filepath: